Crate pegassas_webapp

Modules

dbapi

Structs

LoginForm
User login information

ProbeForm
Probe request frequency information

RegistrationForm

ViewForm
View configuration update form information

Functions

check_username

is_signed_in
Check active user in current session

login
Check user login input against records present in database and if it match start a new user session

main

pass_verify
Verify password string input against the encrypted password held in database

probe_config
Change request interval for a specific probe

registration

request_probe_config
Request probe configuration from database and return it as a Json HttpResponse

setup_logger

to_home
Direct to homepage

view_config
Update view configuration setting in database based on the web form submitted by user

view_config_request
Request view setting configuration from database and return it in a Json Http response