Module pegassas_webapp::dbapi

Manage communication between web front end and server

Structs

BatteryConditionAttribute
Battery table conditions

BatteryRequestQuery
Structure to specify which battery information are required from database

CellConditionAttribute
Cell table condition

CellRequestQuery
Structure to specify which cell information are required from database

GridConditionAttribute
Grid table conditions

GridRequestQuery
Structure to specify which grid information are required from database

HouseConditionAttribute
House_power table conditions

HouseRequestQuery
Structure to specify which House power information are required from database

InverterConditionAttribute
Inverter table conditions

InverterRequestQuery
Structure to specify which inverter information are required from database

TimeSeriesBounds
Time range settings

Functions

clause
Specify time bound for visualisation graph in webpage

request_battery
Http request to server database based on boolean values in BatteryRequestQuery structure

request_cell
Http request to server database based on boolean values in CellRequestQuery structure

request_data
Request data from database and return it in a Json format

request_grid
Http request to server database based on boolean values in GridConditionAttribute structure

request_house
Http request to server database based on boolean values in HouseConditionAttribute structure

request_inverter
Http request to server database based on boolean values in InverterConditionAttribute structure

time_bounds
Specify time bound for visualisation graph in webpage